Adding a small confirm dialog before memo deletes all messages

+++ memo.c
@@ -38,6 +38,7 @@
 
 #include <stdio.h>
 #include <stdlib.h>
+#include <stdbool.h>
 #include <string.h>
 #include <time.h>
 #include <unistd.h>
@@ -78,6 +79,7 @@ FILE *get_memo_file_ptr();
 void  usage();
 void  fail(FILE *out, const char *fmt, ...);
 int   delete_all();
+bool   confirm();  // using bool since the project is built on c99
 void  show_current_memo_file_path();
 NoteStatus_t get_note_status_from_line(const char *line);
 int   mark_note_status(NoteStatus_t status, int id);
@@ -733,6 +735,18 @@ void show_latest(int n)
 		fail(stderr,"%s: counting lines failed\n", __func__);
 }
 
+/* Confirms before deleting all notes
+ * Returns 0 if user-prompt is yes(y/Y), -1 if no.
+ */
+bool confirm()
+{
+    printf("Confirm delete all notes (y/Y)");
+    char confirm = getchar();
+    if (confirm == 'y' || confirm == 'Y')
+        return true;
+    else
+        return false;
+}
 
 /* Deletes all notes. Function actually
  * simply removes .memo file.
@@ -1181,7 +1195,10 @@ int main(int argc, char *argv[])
 			delete_note(atoi(optarg));
 			break;
 		case 'D':
-			delete_all();
+            if (confirm())
+    			delete_all();
+            else
+                printf("Aborting..\n");
 			break;
 		case 'e':
 			export_html(optarg);
